Replace deprecated cgi.parse_header
cgi.parse_header was used to parse the filename param.
Its main features are splitting at ; and unescaping \ and ", but that
has no value in practice. E.g. browsers escape `"` as `%22`, not `\"`.

This patch replaces the use of the deprecated `cgi.parse_header` method
with simple string splitting.

Example:
- Input filename : `_%_%22_& #|<>%20"end`
- Output filename: `_%_%22_& #|<>%20%22end` (%22 is encoded by Firefox)

Simulated with curl:
- Server: `./quick-serve-out.py 8080 --allow-http`
- Client: `curl http://localhost:8080 -X POST -H 'Content-Type: multipart/form-data; boundary=x' -d$'--x\r\nContent-Disposition: form-data; name="file"; filename="_%_%22_& #|<>%20"end"\r\nContent-Type: application/octet-stream\r\n\r\ndata here\r\n--x--\r\n'`
- Output filename: `_%_%22_& #|<>%20"end`